requirements.txt
# ============================================================================
# LLM Module - Core Dependencies
# ============================================================================

# Configuration & Environment
python-dotenv>=1.0.0        # Load API keys from .env file

# LLM Providers
openai>=1.12.0              # OpenAI API (GPT-3.5, GPT-4, GPT-4-turbo)
anthropic>=0.18.0           # Anthropic API (Claude 3 models)

# HTTP & API
requests>=2.31.0            # HTTP requests for API calls
httpx>=0.25.0               # Alternative async HTTP client

# ============================================================================
# Testing
# ============================================================================
pytest>=7.4.0               # Testing framework
pytest-cov>=4.1.0           # Test coverage reports
pytest-asyncio>=0.21.0      # Async test support

# ============================================================================
# Code Quality (Optional but Recommended)
# ============================================================================
black>=23.0.0               # Code formatter
flake8>=6.0.0               # Linter
mypy>=1.0.0                 # Type checking